Switch to optional apply system and finalize extension values on read
